Local and Remote tests implemented and documented
Decided to do a remote test on both a different JVM and the same JVM running the bean.  Results are interesting.  Thought the System.out.println for each line might be causing the slow down, but they are really not.  Comment it out in the remote test and see for yourself.