Reorg and let's have all tests hit the same bean
Move all the test stuff under pref-tests for a little more clarity.  Expose the RESTEasy as a local and remote bean.  The intent here is redo the local and remote RMI tests to hit that bean.  Just trying to reduce the differences between tests.  Also make the REST end point produce a string instead of JSON.  If needed, I'll add another end point that produces JSON, but I don't want serialization to and from JSON to be blamed for the differences in performance at this point.